+++ src/bolt/index.ts
@@ -1,4 +1,4 @@
 export * as customMiddleware from "./customMiddleware";
 export * as interfaces from "./interface";
 export * as blocks from "./block";
-export * as events from "./events";
+export * as events from "./messageEvents";
